MAJOR UPDATE: Converted program to Node.js server
I converted the program to a Node.js server. When you update on your computer, npm install all the dependencies (hopefully you have npm, if not shame on you and get it), and run node main.js. You should then be able to see it at localhost:3000. Note: It only accepts the jpgs as of right now. We will have to implement pdf support later. Note 2: Github doesn't support files larger than 100mb so you will have to download 'values.json' from the google drive and put it in the src/ folder for the program to work.