Algorithm Upgrades - Not to be merged
I've been playing with the algorithm. There use to be a few compounds in the DB that it had problems with, especially 110-85-0_b and 120-72-9_a. Now I don't think we have any that it doesn't find. Although it takes a little longer to run now.

Peak now sorts all peaks from tallest to shortest and then the comparison matches peaks together giving more weight to the taller ones. The other thing I did was start using a basic sorting algorithm with the 3 methods. Instead of just adding all the values together it sorts each methods results and adds a compound to the main list when it  has shown up in 2 of the 3 when iterating through the lists from the top down. Sort of a majorities rule where compounds get added based on how good of a match each thinks it is.